extract logrepl
