+++ logrepl/subscription.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cdc
+package logrepl
 
 import (
 	"context"
@@ -43,6 +43,7 @@ type Subscription struct {
 
 	connConfig pgconn.Config
 	conn       *pgconn.PgConn
+
 	maxWal     uint64
 	walRetain  uint64
 	walFlushed uint64
